+++ examples/sas_run.sh
@@ -5,7 +5,6 @@ dist-s1 run_sas_prep --mgrs_tile_id '11SLT' \
     --memory_strategy 'high' \
     --moderate_confidence_threshold 3.5 \
     --high_confidence_threshold 5.5 \
-    --post_date_buffer_days 1 \
     --apply_water_mask true \
     --device 'cpu' \
     --product_dst_dir '../notebooks/los-angeles' \