+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - Also allows for serialization of algorithm parameters to serparate file as well.
 - Validation for `model_dtype` with `device` (only `bfloat16` is permitted with `GPU`)
 - Validation for external model usage
+- Description of Runconfig Variables
 
 ## Fixed
 - All the tests with the updates above.